pyproject.toml
[project]
name = "report-generator"
version = "0.1.0"
description = ""
authors = [
    {name = "Vyacheslav",email = "slava091103@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-pptx (>=1.0.2,<2.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pillow (>=11.3.0,<12.0.0)"
]

[tool.poetry]
packages = [{include = "report_generator", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
